data_processing/requirements.txt
nibabel~=5.1.0
numpy~=1.24.4
monai~=1.2.0
pandas~=2.0.3
opencv-python~=4.8.0.76
scikit-image~=0.19.3
pydicom~=2.4.3
scipy~=1.10.1
pillow~=9.4.0
